Extract relational metadata and helpers
