+++ crates/truthdb-core/src/rel/mod.rs
@@ -12,8 +12,11 @@ mod api;
 mod batch;
 pub mod collation;
 mod context;
+mod describe;
 mod hash;
+mod helpers;
 mod plan;
+mod sys_views;
 mod value;
 
 pub use api::{
@@ -22,11 +25,16 @@ pub use api::{
 };
 pub use batch::{execute_batch, execute_batch_streamed, execute_batch_with_params};
 pub use context::{Isolation, TxnContext};
+pub use describe::describe_first_result_set;
+pub use helpers::{SqlStatement, render_cell};
 
 use api::{DeferredDone, DiscardEmitter, XACT_ABORT_SEVERITY};
 use batch::BatchRun;
 use batch::done_command;
 use context::TableVar;
+use describe::produces_rowset;
+use helpers::*;
+use sys_views::*;
 
 use truthdb_sql::ast::{
     AlterAction, AlterDatabase, AlterTable, CheckConstraint, ColumnDef, CreateFunction,
@@ -1370,160 +1378,6 @@ fn run_block(
         }
     }
     Ok(Flow::Normal)
-}
-
-/// `sp_describe_first_result_set`: the column metadata of `tsql`'s first
-/// statement's result set, without executing it. Covers exactly the shapes
-/// whose output columns are statically derivable — the single-table SELECTs
-/// [`scan_plan`] recognises (their COLMETADATA is derived before the first
-/// row at execution too). A statement producing no result set describes as
-/// zero rows, matching SQL Server. Everything else — joins, aggregates,
-/// computed columns — has result types this engine only learns by executing
-/// (`infer_type` over the values), so it answers 11514: honest, and the same
-/// static-type-derivation gap Stage 8's output streaming tracks.
-///
-/// `@tsql`'s declared parameters need no binding here: the planner treats an
-/// unresolvable variable as a non-seekable predicate and falls back to the
-/// scan, and output columns come from the table schema either way — pinned by
-/// the parameterized-describe test.
-pub fn describe_first_result_set(storage: &Storage, tsql: &str) -> Result<RowSet, SqlError> {
-    let undeterminable = || {
-        SqlError::new(
-            11514,
-            16,
-            1,
-            "The metadata could not be determined because the statement's result-set \
-             types are not statically derivable by this server (single-table SELECTs \
-             are; joins, aggregates and computed columns are typed at execution).",
-        )
-    };
-    let statements = truthdb_sql::parse(tsql)?;
-    // The contract is the batch's first RESULT SET, not its first statement:
-    // `INSERT ...; SELECT ...` describes the SELECT. TRY blocks are entered
-    // (their statements run unless an error preempts them); a rowset only a
-    // CATCH can produce stays undescribed — that path is conditional.
-    let columns = match first_rowset_statement(&statements) {
-        None => Vec::new(),
-        Some(Statement::Select(select)) => {
-            // TOP never changes the columns, and `scan_plan` rejects `TOP 0`
-            // for an execution-path reason describe does not share.
-            let mut select = select.clone();
-            select.top = None;
-            let mut ctx = TxnContext::default();
-            // Describe derives column metadata without executing or reading data;
-            // it does not enforce object permissions (and its throwaway context
-            // carries no session identity), so the shared `scan_plan` must not
-            // treat it as a denied read.
-            ctx.security.bypass = true;
-            match scan_plan(storage, &select, &ctx.eval_context()) {
-                Some(plan) => plan.columns,
-                None => return Err(undeterminable()),
-            }
-        }
-        Some(_) => return Err(undeterminable()),
-    };
-
-    let mut rows = Vec::new();
-    for (ordinal, column) in columns.iter().enumerate() {
-        let (type_id, type_name) = describe_type(&column.column_type);
-        rows.push(vec![
-            Datum::Bit(false),                    // is_hidden
-            Datum::Int(ordinal as i32 + 1),       // column_ordinal
-            Datum::NVarChar(column.name.clone()), // name
-            Datum::Bit(true), // is_nullable (result metadata is nullable, as at execution)
-            Datum::Int(type_id), // system_type_id
-            Datum::NVarChar(type_name), // system_type_name
-        ]);
-    }
-    Ok(RowSet {
-        columns: vec![
-            ResultColumn {
-                name: "is_hidden".into(),
-                column_type: ColumnType::Bit,
-            },
-            ResultColumn {
-                name: "column_ordinal".into(),
-                column_type: ColumnType::Int,
-            },
-            ResultColumn {
-                name: "name".into(),
-                column_type: ColumnType::NVarChar { max_len: 128 },
-            },
-            ResultColumn {
-                name: "is_nullable".into(),
-                column_type: ColumnType::Bit,
-            },
-            ResultColumn {
-                name: "system_type_id".into(),
-                column_type: ColumnType::Int,
-            },
-            ResultColumn {
-                name: "system_type_name".into(),
-                column_type: ColumnType::NVarChar { max_len: 256 },
-            },
-        ],
-        rows,
-    })
-}
-
-/// The first statement in `statements` that opens a result set, descending
-/// into TRY blocks (entered unless an error preempts them) but not CATCH
-/// blocks (conditional).
-fn first_rowset_statement(statements: &[Statement]) -> Option<&Statement> {
-    statements.iter().find_map(|statement| match statement {
-        Statement::TryCatch { try_block, .. } => first_rowset_statement(try_block),
-        s if produces_rowset(s) => Some(s),
-        _ => None,
-    })
-}
-
-/// A column type's `sys.types` id and its `system_type_name` spelling, as
-/// `sp_describe_first_result_set` reports them.
-fn describe_type(column_type: &ColumnType) -> (i32, String) {
-    match column_type {
-        ColumnType::TinyInt => (48, "tinyint".into()),
-        ColumnType::SmallInt => (52, "smallint".into()),
-        ColumnType::Int => (56, "int".into()),
-        ColumnType::BigInt => (127, "bigint".into()),
-        ColumnType::Bit => (104, "bit".into()),
-        ColumnType::Real => (59, "real".into()),
-        ColumnType::Float => (62, "float".into()),
-        ColumnType::Decimal { precision, scale } => (106, format!("decimal({precision},{scale})")),
-        ColumnType::Date => (40, "date".into()),
-        ColumnType::Time => (41, "time".into()),
-        ColumnType::DateTime2 => (42, "datetime2".into()),
-        ColumnType::UniqueIdentifier => (36, "uniqueidentifier".into()),
-        ColumnType::VarChar { max_len } => (167, format!("varchar({max_len})")),
-        ColumnType::NVarChar { max_len } => (231, format!("nvarchar({max_len})")),
-        ColumnType::VarBinary { max_len } => (165, format!("varbinary({max_len})")),
-        ColumnType::VarCharMax => (167, "varchar(max)".into()),
-        ColumnType::NVarCharMax => (231, "nvarchar(max)".into()),
-        ColumnType::VarBinaryMax => (165, "varbinary(max)".into()),
-    }
-}
-
-/// Whether a statement can open a result set on the stream: a `SELECT` that
-/// returns rows (an assignment `SELECT @v = ...` returns none). `SET
-/// SHOWPLAN_TEXT`'s plan rows ride the same `SELECT` arm.
-fn produces_rowset(statement: &Statement) -> bool {
-    match statement {
-        Statement::Select(select) => !select
-            .items
-            .iter()
-            .any(|i| matches!(i, SelectItem::Assign { .. })),
-        // EXEC's inner batch — and any control-flow body — may open result
-        // sets; conservative.
-        Statement::Exec(_) => true,
-        Statement::Block { .. } | Statement::If { .. } | Statement::While { .. } => true,
-        // The header/file-list restore verbs return a rowset; VERIFYONLY and the
-        // offline DATABASE/LOG forms do not.
-        Statement::Restore { mode, .. } => {
-            matches!(mode, RestoreMode::HeaderOnly | RestoreMode::FileListOnly)
-        }
-        // A `FETCH` without an INTO list returns the fetched row to the client.
-        Statement::FetchCursor { into, .. } => into.is_empty(),
-        _ => false,
-    }
 }
 
 /// One executed statement's outcome, from [`exec_statement_streamed`].
@@ -12955,1483 +12809,4 @@ fn hash_join(
         }
     }
     Ok(())
-}
-
-// ---- sys.* virtual sources ---------------------------------------------
-
-fn nvarchar(name: &str, max_len: u16) -> ResultColumn {
-    ResultColumn {
-        name: name.to_string(),
-        column_type: ColumnType::NVarChar { max_len },
-    }
-}
-
-fn int_col(name: &str) -> ResultColumn {
-    ResultColumn {
-        name: name.to_string(),
-        column_type: ColumnType::Int,
-    }
-}
-
-fn bigint_col(name: &str) -> ResultColumn {
-    ResultColumn {
-        name: name.to_string(),
-        column_type: ColumnType::BigInt,
-    }
-}
-
-fn bit_col(name: &str) -> ResultColumn {
-    ResultColumn {
-        name: name.to_string(),
-        column_type: ColumnType::Bit,
-    }
-}
-
-/// `RESTORE VERIFYONLY/HEADERONLY/FILELISTONLY` (online, read-only), plus a clear
-/// error for the offline-only `RESTORE DATABASE`/`LOG`. Like BACKUP, restore is
-/// illegal inside a transaction (3021).
-fn exec_restore(
-    mode: RestoreMode,
-    path: &str,
-    txn_ctx: &TxnContext,
-) -> Result<StatementResult, SqlError> {
-    if txn_ctx.in_txn() {
-        return Err(SqlError::new(
-            3021,
-            16,
-            1,
-            "Cannot perform a backup or restore operation within a transaction.".to_string(),
-        ));
-    }
-    let file = std::path::Path::new(path);
-    let terminating = |verb: &str, e: std::io::Error| {
-        SqlError::new(
-            3013,
-            16,
-            1,
-            format!("RESTORE {verb} is terminating abnormally. {e}"),
-        )
-    };
-    match mode {
-        RestoreMode::VerifyOnly => {
-            crate::backup::verify(file).map_err(|e| terminating("VERIFYONLY", e))?;
-            Ok(StatementResult::Done)
-        }
-        RestoreMode::HeaderOnly => {
-            let header =
-                crate::backup::read_header(file).map_err(|e| terminating("HEADERONLY", e))?;
-            Ok(StatementResult::Rows(restore_headeronly_rows(&header)))
-        }
-        RestoreMode::FileListOnly => {
-            let header =
-                crate::backup::read_header(file).map_err(|e| terminating("FILELISTONLY", e))?;
-            Ok(StatementResult::Rows(restore_filelist_rows(&header)))
-        }
-        RestoreMode::Database | RestoreMode::Log => Err(SqlError::new(
-            3101,
-            16,
-            1,
-            "Exclusive access could not be obtained because the database is in use. TruthDB \
-             restores a database offline: stop the server and run `truthdb-cli restore`."
-                .to_string(),
-        )),
-    }
-}
-
-/// One metadata row for `RESTORE HEADERONLY`.
-fn restore_headeronly_rows(header: &crate::backup::BackupHeader) -> RowSet {
-    let columns = vec![
-        int_col("BackupType"),
-        int_col("FormatVersion"),
-        int_col("PageSize"),
-        bigint_col("DatabaseSize"),
-        bit_col("Checksum"),
-        bit_col("CopyOnly"),
-        bigint_col("RedoStartLSN"),
-        bigint_col("LastCommittedSeq"),
-        int_col("DbOptions"),
-        nvarchar("Collation", 128),
-        bigint_col("BackupFinishDate"),
-    ];
-    // BackupType: 1 = full database, 2 = transaction log (SQL Server's coding).
-    let backup_type = if header.flags.log_backup { 2 } else { 1 };
-    let row = vec![
-        Datum::Int(backup_type),
-        Datum::Int(header.format_version as i32),
-        Datum::Int(header.page_size as i32),
-        Datum::BigInt(header.total_size as i64),
-        Datum::Bit(header.flags.checksum),
-        Datum::Bit(header.flags.copy_only),
-        Datum::BigInt(header.redo_start_lsn as i64),
-        Datum::BigInt(header.last_committed_seq as i64),
-        Datum::Int(header.db_options as i32),
-        match &header.default_collation {
-            Some(c) => Datum::NVarChar(c.clone()),
-            None => Datum::Null,
-        },
-        Datum::BigInt(header.finished_at_millis as i64),
-    ];
-    RowSet {
-        columns,
-        rows: vec![row],
-    }
-}
-
-/// The data + log "files" for `RESTORE FILELISTONLY`. A log-only archive reports
-/// a zero data size (it holds no page data) but keeps its WAL region size;
-/// `HEADERONLY`'s BackupType = 2 marks it as a log backup.
-fn restore_filelist_rows(header: &crate::backup::BackupHeader) -> RowSet {
-    let columns = vec![
-        nvarchar("LogicalName", 128),
-        nvarchar("Type", 1),
-        bigint_col("Size"),
-    ];
-    let rows = vec![
-        vec![
-            Datum::NVarChar("truthdb_data".to_string()),
-            Datum::NVarChar("D".to_string()),
-            Datum::BigInt(header.data_size as i64),
-        ],
-        vec![
-            Datum::NVarChar("truthdb_log".to_string()),
-            Datum::NVarChar("L".to_string()),
-            Datum::BigInt(header.wal_size as i64),
-        ],
-    ];
-    RowSet { columns, rows }
-}
-
-fn sys_tables(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![int_col("object_id"), nvarchar("name", 128)];
-    let rows = storage
-        .rel_tables()
-        .into_iter()
-        .filter(|def| def.database_id == db_id)
-        // Only base tables. (The `!is_view()` filter alone let procedures leak
-        // in — a pre-existing gap — so exclude every non-table object kind.)
-        .filter(|def| {
-            !def.is_view() && !def.is_procedure() && !def.is_function() && !def.is_trigger()
-        })
-        .map(|def| vec![Datum::Int(def.object_id as i32), Datum::NVarChar(def.name)])
-        .collect();
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-/// `sys.views` — one row per view, with its stored definition text.
-/// `sys.dm_repl_slots` (Stage 18): the primary's replication slots and the
-/// WAL-ring positions they pin. Empty on a standby (and on a primary with no
-/// standby ever connected).
-fn sys_dm_repl_slots(storage: &Storage) -> Source {
-    let columns = vec![
-        int_col("slot_id"),
-        bigint_col("held_lsn"),
-        bigint_col("wal_head"),
-        bigint_col("wal_tail"),
-        bigint_col("retained_bytes"),
-    ];
-    let head = storage.wal_head();
-    let tail = storage.wal_tail();
-    let rows: Vec<Vec<Datum>> = storage
-        .repl_slots_snapshot()
-        .into_iter()
-        .map(|(id, lsn)| {
-            vec![
-                Datum::Int(id as i32),
-                Datum::BigInt(lsn as i64),
-                Datum::BigInt(head as i64),
-                Datum::BigInt(tail as i64),
-                Datum::BigInt(tail.saturating_sub(lsn) as i64),
-            ]
-        })
-        .collect();
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-/// `sys.dm_repl_replica_states` (Stage 18): this node's replication role and,
-/// on a primary, one row per replication slot with its connection and lag
-/// state; on a standby, one row describing its own applied position.
-fn sys_dm_repl_replica_states(storage: &Storage) -> Source {
-    let columns = vec![
-        nvarchar("role", 20),
-        int_col("node_id"),
-        ResultColumn {
-            name: "is_connected".into(),
-            column_type: ColumnType::Bit,
-        },
-        bigint_col("acked_lsn"),
-        bigint_col("durable_lsn"),
-        bigint_col("lag_bytes"),
-        nvarchar("sync_state", 30),
-        bigint_col("epoch"),
-    ];
-    let epoch = storage.epoch() as i64;
-    let rows = if storage.is_standby() {
-        vec![vec![
-            Datum::NVarChar("STANDBY".into()),
-            Datum::Null,
-            Datum::Null,
-            Datum::BigInt(storage.applied_lsn() as i64),
-            Datum::BigInt(storage.wal_flushed_lsn() as i64),
-            Datum::BigInt(0),
-            Datum::NVarChar("NOT_APPLICABLE".into()),
-            Datum::BigInt(epoch),
-        ]]
-    } else {
-        let durable = storage.wal_flushed_lsn();
-        let connected = storage.repl_connected_nodes();
-        let sync_state = match storage.sync_commit_status() {
-            None => "ASYNC",
-            Some(false) => "SYNCHRONIZED",
-            Some(true) => "NOT_SYNCHRONIZED",
-        };
-        storage
-            .repl_slots_snapshot()
-            .into_iter()
-            .map(|(id, lsn)| {
-                vec![
-                    Datum::NVarChar("PRIMARY".into()),
-                    Datum::Int(id as i32),
-                    Datum::Bit(connected.contains(&id)),
-                    Datum::BigInt(lsn as i64),
-                    Datum::BigInt(durable as i64),
-                    Datum::BigInt(durable.saturating_sub(lsn) as i64),
-                    Datum::NVarChar(sync_state.into()),
-                    Datum::BigInt(epoch),
-                ]
-            })
-            .collect()
-    };
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-/// `sys.databases` (Stage 14, SSMS query-window probes): the one database
-/// this instance serves, with the columns tools actually read. The
-/// versioning flags report the live `ALTER DATABASE` options.
-fn sys_databases(storage: &Storage) -> Source {
-    let columns = vec![
-        nvarchar("name", 128),
-        int_col("database_id"),
-        int_col("compatibility_level"),
-        nvarchar("collation_name", 128),
-        nvarchar("user_access_desc", 60),
-        nvarchar("state_desc", 60),
-        nvarchar("recovery_model_desc", 60),
-        int_col("snapshot_isolation_state"),
-        ResultColumn {
-            name: "is_read_committed_snapshot_on".into(),
-            column_type: ColumnType::Bit,
-        },
-        ResultColumn {
-            name: "is_read_only".into(),
-            column_type: ColumnType::Bit,
-        },
-    ];
-    // One row per database. The option columns are instance-wide (one
-    // shared log and version store), so every row reports the same values —
-    // the documented level-1 deviation.
-    let rows = storage
-        .rel_databases()
-        .into_iter()
-        .map(|(id, name)| {
-            vec![
-                Datum::NVarChar(name),
-                Datum::Int(id as i32),
-                Datum::Int(160),
-                Datum::NVarChar("SQL_Latin1_General_CP1_CI_AS".into()),
-                Datum::NVarChar("MULTI_USER".into()),
-                Datum::NVarChar("ONLINE".into()),
-                Datum::NVarChar(
-                    if storage.recovery_model_full() {
-                        "FULL"
-                    } else {
-                        "SIMPLE"
-                    }
-                    .into(),
-                ),
-                Datum::Int(storage.snapshot_isolation_allowed() as i32),
-                Datum::Bit(storage.rcsi_enabled()),
-                Datum::Bit(false),
-            ]
-        })
-        .collect();
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-/// `sys.configurations` (Stage 14): the handful of static rows connection
-/// tools probe. Values are INT here (SQL Server uses sql_variant) — a
-/// documented simplification.
-fn sys_configurations() -> Source {
-    let columns = vec![
-        int_col("configuration_id"),
-        nvarchar("name", 35),
-        int_col("value"),
-        int_col("minimum"),
-        int_col("maximum"),
-        int_col("value_in_use"),
-        nvarchar("description", 255),
-        ResultColumn {
-            name: "is_dynamic".into(),
-            column_type: ColumnType::Bit,
-        },
-        ResultColumn {
-            name: "is_advanced".into(),
-            column_type: ColumnType::Bit,
-        },
-    ];
-    let entry =
-        |id: i32, name: &str, value: i32, min: i32, max: i32, dynamic: bool, advanced: bool| {
-            vec![
-                Datum::Int(id),
-                Datum::NVarChar(name.into()),
-                Datum::Int(value),
-                Datum::Int(min),
-                Datum::Int(max),
-                Datum::Int(value),
-                Datum::NVarChar(name.into()),
-                Datum::Bit(dynamic),
-                Datum::Bit(advanced),
-            ]
-        };
-    let rows = vec![
-        entry(16384, "show advanced options", 0, 0, 1, true, false),
-        entry(1539, "user options", 0, 0, 32767, true, false),
-        entry(
-            1544,
-            "max server memory (MB)",
-            i32::MAX,
-            16,
-            i32::MAX,
-            true,
-            true,
-        ),
-    ];
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_views(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        int_col("object_id"),
-        nvarchar("name", 128),
-        nvarchar("definition", 4000),
-    ];
-    let rows = storage
-        .rel_tables()
-        .into_iter()
-        .filter(|def| def.database_id == db_id)
-        .filter_map(|def| {
-            def.view_query.map(|q| {
-                vec![
-                    Datum::Int(def.object_id as i32),
-                    Datum::NVarChar(def.name),
-                    Datum::NVarChar(q),
-                ]
-            })
-        })
-        .collect();
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-/// `sys.sql_modules`: the SQL definition of each module (currently views), keyed
-/// by `object_id`. SQL Server surfaces view/procedure/trigger text here; today
-/// only views carry a definition.
-fn sys_sql_modules(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![int_col("object_id"), nvarchar("definition", 4000)];
-    let rows = storage
-        .rel_tables()
-        .into_iter()
-        .filter(|def| def.database_id == db_id)
-        .filter_map(|def| {
-            // Views store their SELECT; procedures and functions their body.
-            let definition = def
-                .view_query
-                .clone()
-                .or_else(|| def.procedure.as_ref().map(|p| p.body.clone()))
-                .or_else(|| {
-                    def.function.as_ref().map(|f| match &f.returns {
-                        FunctionReturns::Scalar { body, .. } => body.clone(),
-                        FunctionReturns::InlineTable { select_text } => select_text.clone(),
-                        FunctionReturns::MultiStatementTable { body, .. } => body.clone(),
-                    })
-                })
-                .or_else(|| def.trigger.as_ref().map(|t| t.body.clone()))?;
-            Some(vec![
-                Datum::Int(def.object_id as i32),
-                Datum::NVarChar(definition),
-            ])
-        })
-        .collect();
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_procedures(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![nvarchar("name", 128), int_col("object_id")];
-    let rows = storage
-        .rel_tables()
-        .into_iter()
-        .filter(|def| def.database_id == db_id)
-        .filter(|def| def.is_procedure())
-        .map(|def| {
-            vec![
-                Datum::NVarChar(def.name.clone()),
-                Datum::Int(def.object_id as i32),
-            ]
-        })
-        .collect();
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_triggers(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        nvarchar("name", 128),
-        int_col("object_id"),
-        int_col("parent_id"),
-        nvarchar("type", 2),
-        int_col("is_disabled"),
-        int_col("is_instead_of_trigger"),
-    ];
-    let rows = storage
-        .rel_tables()
-        .into_iter()
-        .filter(|def| def.database_id == db_id)
-        .filter_map(|def| {
-            let trigger = def.trigger.as_ref()?;
-            Some(vec![
-                Datum::NVarChar(def.name.clone()),
-                Datum::Int(def.object_id as i32),
-                Datum::Int(trigger.parent_object_id as i32),
-                Datum::NVarChar("TR".to_string()),
-                Datum::Int(i32::from(trigger.is_disabled)),
-                Datum::Int(i32::from(trigger.is_instead_of)),
-            ])
-        })
-        .collect();
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_trigger_events(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        int_col("object_id"),
-        int_col("type"),
-        nvarchar("type_desc", 128),
-    ];
-    let mut rows = Vec::new();
-    for def in storage.rel_tables() {
-        if def.database_id != db_id {
-            continue;
-        }
-        let Some(trigger) = def.trigger.as_ref() else {
-            continue;
-        };
-        for event in &trigger.events {
-            let (code, desc) = match event {
-                catalog::TriggerEvent::Insert => (1, "INSERT"),
-                catalog::TriggerEvent::Update => (2, "UPDATE"),
-                catalog::TriggerEvent::Delete => (3, "DELETE"),
-            };
-            rows.push(vec![
-                Datum::Int(def.object_id as i32),
-                Datum::Int(code),
-                Datum::NVarChar(desc.to_string()),
-            ]);
-        }
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_server_principals(storage: &Storage) -> Source {
-    let columns = vec![
-        nvarchar("name", 128),
-        int_col("principal_id"),
-        nvarchar("type", 1),
-        nvarchar("type_desc", 60),
-        int_col("is_disabled"),
-    ];
-    let mut rows: Vec<Vec<Datum>> = storage
-        .rel_logins()
-        .into_iter()
-        .filter_map(|def| {
-            let principal = def.principal.as_ref()?;
-            Some(vec![
-                Datum::NVarChar(def.name.clone()),
-                Datum::Int(def.object_id as i32),
-                // SQL logins: type 'S' / SQL_LOGIN.
-                Datum::NVarChar("S".to_string()),
-                Datum::NVarChar("SQL_LOGIN".to_string()),
-                Datum::Int(i32::from(principal.is_disabled)),
-            ])
-        })
-        .collect();
-    // The fixed server roles (today: sysadmin) — type 'R' / SERVER_ROLE.
-    for fixed in crate::storage::FIXED_PRINCIPALS
-        .iter()
-        .filter(|p| p.is_server)
-    {
-        rows.push(vec![
-            Datum::NVarChar(fixed.name.to_string()),
-            Datum::Int(fixed.id as i32),
-            Datum::NVarChar("R".to_string()),
-            Datum::NVarChar("SERVER_ROLE".to_string()),
-            Datum::Int(0),
-        ]);
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_sql_logins(storage: &Storage) -> Source {
-    let columns = vec![
-        nvarchar("name", 128),
-        int_col("principal_id"),
-        int_col("is_disabled"),
-    ];
-    let rows = storage
-        .rel_logins()
-        .into_iter()
-        .filter_map(|def| {
-            let principal = def.principal.as_ref()?;
-            Some(vec![
-                Datum::NVarChar(def.name.clone()),
-                Datum::Int(def.object_id as i32),
-                Datum::Int(i32::from(principal.is_disabled)),
-            ])
-        })
-        .collect();
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_database_principals(storage: &Storage) -> Source {
-    use crate::relstore::catalog::PrincipalKind;
-    let columns = vec![
-        nvarchar("name", 128),
-        int_col("principal_id"),
-        nvarchar("type", 1),
-        nvarchar("type_desc", 60),
-        nvarchar("default_schema_name", 128),
-        int_col("owning_principal_id"),
-    ];
-    // A user (SQL_USER 'S') defaults to the dbo schema; a role (DATABASE_ROLE
-    // 'R') has no default schema.
-    let row = |name: String, id: u32, kind: PrincipalKind| {
-        let (type_code, type_desc) = match kind {
-            PrincipalKind::Role => ("R", "DATABASE_ROLE"),
-            _ => ("S", "SQL_USER"),
-        };
-        let default_schema = if matches!(kind, PrincipalKind::User) {
-            Datum::NVarChar("dbo".to_string())
-        } else {
-            Datum::Null
-        };
-        vec![
-            Datum::NVarChar(name),
-            Datum::Int(id as i32),
-            Datum::NVarChar(type_code.to_string()),
-            Datum::NVarChar(type_desc.to_string()),
-            default_schema,
-            Datum::Null, // owning_principal_id
-        ]
-    };
-    let mut rows: Vec<Vec<Datum>> = Vec::new();
-    // Fixed database principals (dbo + the fixed database roles + public); the
-    // server-scoped sysadmin role belongs to sys.server_principals instead.
-    for fixed in crate::storage::FIXED_PRINCIPALS
-        .iter()
-        .filter(|p| !p.is_server)
-    {
-        rows.push(row(fixed.name.to_string(), fixed.id, fixed.kind));
-    }
-    for def in storage.rel_database_principals() {
-        if let Some(principal) = def.principal.as_ref() {
-            rows.push(row(def.name.clone(), def.object_id, principal.kind));
-        }
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_database_role_members(storage: &Storage) -> Source {
-    let columns = vec![int_col("role_principal_id"), int_col("member_principal_id")];
-    let mut rows: Vec<Vec<Datum>> = Vec::new();
-    // Synthesized: the dbo user is a member of db_owner.
-    rows.push(vec![
-        Datum::Int(crate::storage::DB_OWNER_ID as i32),
-        Datum::Int(crate::storage::DBO_ID as i32),
-    ]);
-    // Stored database membership edges (member -> role, from each member's row).
-    for def in storage.rel_database_principals() {
-        if let Some(principal) = def.principal.as_ref() {
-            for &role_id in &principal.member_of {
-                rows.push(vec![
-                    Datum::Int(role_id as i32),
-                    Datum::Int(def.object_id as i32),
-                ]);
-            }
-        }
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_database_permissions(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        int_col("class"),
-        nvarchar("class_desc", 60),
-        int_col("major_id"),
-        int_col("minor_id"),
-        int_col("grantee_principal_id"),
-        nvarchar("permission_name", 128),
-        nvarchar("state", 1),
-        nvarchar("state_desc", 60),
-    ];
-    let mut rows: Vec<Vec<Datum>> = Vec::new();
-    for def in storage.rel_tables() {
-        if def.database_id != db_id {
-            continue;
-        }
-        for perm in &def.permissions {
-            let (state, state_desc) = if perm.deny {
-                ("D", "DENY")
-            } else {
-                ("G", "GRANT")
-            };
-            rows.push(vec![
-                Datum::Int(1), // class 1 = OBJECT_OR_COLUMN
-                Datum::NVarChar("OBJECT_OR_COLUMN".to_string()),
-                Datum::Int(def.object_id as i32),
-                Datum::Int(0), // minor_id 0 = the whole object (no column-level)
-                Datum::Int(perm.grantee as i32),
-                Datum::NVarChar(perm.action.name().to_string()),
-                Datum::NVarChar(state.to_string()),
-                Datum::NVarChar(state_desc.to_string()),
-            ]);
-        }
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_parameters(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        int_col("object_id"),
-        nvarchar("name", 128),
-        int_col("parameter_id"),
-        nvarchar("system_type_name", 128),
-        int_col("is_output"),
-        int_col("has_default_value"),
-    ];
-    let mut rows = Vec::new();
-    let mut push_param = |object_id: u32,
-                          name: String,
-                          id: i32,
-                          type_spec: String,
-                          output: bool,
-                          has_default: bool| {
-        rows.push(vec![
-            Datum::Int(object_id as i32),
-            Datum::NVarChar(name),
-            Datum::Int(id),
-            Datum::NVarChar(type_spec),
-            Datum::Int(i32::from(output)),
-            Datum::Int(i32::from(has_default)),
-        ]);
-    };
-    for def in storage.rel_tables() {
-        if def.database_id != db_id {
-            continue;
-        }
-        if let Some(procedure) = &def.procedure {
-            for (index, param) in procedure.params.iter().enumerate() {
-                push_param(
-                    def.object_id,
-                    format!("@{}", param.name),
-                    index as i32 + 1,
-                    param.type_spec.clone(),
-                    param.output,
-                    param.default.is_some(),
-                );
-            }
-        } else if let Some(function) = &def.function {
-            // A SCALAR function's return value is parameter_id 0 (empty name,
-            // is_output set — SQL Server's convention). A table-valued function
-            // returns a table, so it has no scalar return parameter.
-            if let FunctionReturns::Scalar { type_spec, .. } = &function.returns {
-                push_param(
-                    def.object_id,
-                    String::new(),
-                    0,
-                    type_spec.clone(),
-                    true,
-                    false,
-                );
-            }
-            for (index, param) in function.params.iter().enumerate() {
-                push_param(
-                    def.object_id,
-                    format!("@{}", param.name),
-                    index as i32 + 1,
-                    param.type_spec.clone(),
-                    false,
-                    param.default.is_some(),
-                );
-            }
-        }
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_objects(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        nvarchar("name", 128),
-        int_col("object_id"),
-        nvarchar("type", 2),
-        nvarchar("type_desc", 60),
-    ];
-    let rows = storage
-        .rel_tables()
-        .into_iter()
-        .filter(|def| def.database_id == db_id)
-        .map(|def| {
-            // SQL Server's single-letter codes carry a trailing space; the
-            // two-letter function codes fill CHAR(2) exactly.
-            let (code, desc) = if let Some(function) = &def.function {
-                match function.returns {
-                    FunctionReturns::Scalar { .. } => ("FN", "SQL_SCALAR_FUNCTION"),
-                    FunctionReturns::InlineTable { .. } => {
-                        ("IF", "SQL_INLINE_TABLE_VALUED_FUNCTION")
-                    }
-                    FunctionReturns::MultiStatementTable { .. } => {
-                        ("TF", "SQL_TABLE_VALUED_FUNCTION")
-                    }
-                }
-            } else if def.is_procedure() {
-                ("P ", "SQL_STORED_PROCEDURE")
-            } else if def.is_trigger() {
-                ("TR", "SQL_TRIGGER")
-            } else if def.is_view() {
-                ("V ", "VIEW")
-            } else {
-                ("U ", "USER_TABLE")
-            };
-            vec![
-                Datum::NVarChar(def.name.clone()),
-                Datum::Int(def.object_id as i32),
-                Datum::NVarChar(code.to_string()),
-                Datum::NVarChar(desc.to_string()),
-            ]
-        })
-        .collect();
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_columns(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        int_col("object_id"),
-        nvarchar("name", 128),
-        int_col("column_id"),
-        nvarchar("type", 128),
-        ResultColumn {
-            name: "is_nullable".to_string(),
-            column_type: ColumnType::Bit,
-        },
-        nvarchar("collation_name", 128),
-    ];
-    let mut rows = Vec::new();
-    for def in storage.rel_tables() {
-        if def.database_id != db_id {
-            continue;
-        }
-        for (index, (name, type_spec, nullable)) in def.columns.iter().enumerate() {
-            let collation = def
-                .collations
-                .get(index)
-                .cloned()
-                .flatten()
-                .map(Datum::NVarChar)
-                .unwrap_or(Datum::Null);
-            rows.push(vec![
-                Datum::Int(def.object_id as i32),
-                Datum::NVarChar(name.clone()),
-                Datum::Int(index as i32 + 1),
-                Datum::NVarChar(type_spec.clone()),
-                Datum::Bit(*nullable),
-                collation,
-            ]);
-        }
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_indexes(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        int_col("object_id"),
-        int_col("index_id"),
-        nvarchar("name", 128),
-        ResultColumn {
-            name: "is_unique".to_string(),
-            column_type: ColumnType::Bit,
-        },
-    ];
-    let mut rows = Vec::new();
-    for def in storage.rel_tables() {
-        if def.database_id != db_id {
-            continue;
-        }
-        for index in &def.indexes {
-            rows.push(vec![
-                Datum::Int(def.object_id as i32),
-                Datum::Int(index.object_id as i32),
-                Datum::NVarChar(index.name.clone()),
-                Datum::Bit(index.unique),
-            ]);
-        }
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_check_constraints(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        nvarchar("name", 128),
-        int_col("parent_object_id"),
-        nvarchar("definition", 4000),
-    ];
-    let mut rows = Vec::new();
-    for def in storage.rel_tables() {
-        if def.database_id != db_id {
-            continue;
-        }
-        for check in &def.check_constraints {
-            rows.push(vec![
-                Datum::NVarChar(check.name.clone()),
-                Datum::Int(def.object_id as i32),
-                Datum::NVarChar(format!("({})", check.predicate)),
-            ]);
-        }
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_foreign_keys(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        nvarchar("name", 128),
-        int_col("parent_object_id"),
-        int_col("referenced_object_id"),
-    ];
-    // Resolve parent (referenced) table names to object ids.
-    let tables: Vec<TableDef> = storage
-        .rel_tables()
-        .into_iter()
-        .filter(|t| t.database_id == db_id)
-        .collect();
-    let oid_of = |name: &str| {
-        tables
-            .iter()
-            .find(|t| t.name.eq_ignore_ascii_case(name))
-            .map(|t| t.object_id)
-    };
-    let mut rows = Vec::new();
-    for def in &tables {
-        for fk in &def.foreign_keys {
-            rows.push(vec![
-                Datum::NVarChar(fk.name.clone()),
-                Datum::Int(def.object_id as i32),
-                oid_of(&fk.parent)
-                    .map(|o| Datum::Int(o as i32))
-                    .unwrap_or(Datum::Null),
-            ]);
-        }
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-fn sys_default_constraints(storage: &Storage, db_id: u32) -> Source {
-    let columns = vec![
-        nvarchar("name", 128),
-        int_col("parent_object_id"),
-        int_col("parent_column_id"),
-        nvarchar("definition", 4000),
-    ];
-    // Inline column DEFAULTs are unnamed; SQL Server auto-names them
-    // `DF__<table>__<column>__...`. We synthesize a stable `DF__<table>__<col>`.
-    let mut rows = Vec::new();
-    for def in storage.rel_tables() {
-        if def.database_id != db_id {
-            continue;
-        }
-        for (index, text) in def.defaults.iter().enumerate() {
-            let Some(text) = text else { continue };
-            let column = &def.columns[index].0;
-            rows.push(vec![
-                Datum::NVarChar(format!("DF__{}__{}", def.name, column)),
-                Datum::Int(def.object_id as i32),
-                Datum::Int(index as i32 + 1),
-                Datum::NVarChar(format!("({text})")),
-            ]);
-        }
-    }
-    let collations = vec![None; columns.len()];
-    let qualifiers = vec![None; columns.len()];
-    Source {
-        columns,
-        qualifiers,
-        collations,
-        rows: SourceRows::Materialized(rows),
-    }
-}
-
-// ---- helpers ------------------------------------------------------------
-
-/// Evaluates a constant expression (INSERT VALUES): no columns in scope.
-fn eval_constant(expr: &Expr, eval_ctx: &EvalContext) -> Result<SqlValue, SqlError> {
-    let empty: Vec<String> = Vec::new();
-    eval::eval(expr, &[], &empty, eval_ctx)
-}
-
-fn column_index(schema: &Schema, name: &str) -> Option<usize> {
-    schema
-        .columns
-        .iter()
-        .position(|c| c.name.eq_ignore_ascii_case(name))
-}
-
-/// Strips an optional `dbo.` schema prefix (Stage 3 has a single user
-/// schema); `sys.` names are handled separately as catalog views.
-fn strip_schema(name: &str) -> &str {
-    name.split_once('.')
-        .filter(|(schema, _)| schema.eq_ignore_ascii_case("dbo"))
-        .map(|(_, rest)| rest)
-        .unwrap_or(name)
-}
-
-/// Case-insensitive table resolution (single `dbo` schema in Stage 3). An
-/// optional `dbo.` schema prefix is accepted and stripped.
-/// The base-table object ids that a read of `name` must Shared-lock: the table
-/// itself, or — for a view — the base tables its definition reads. `sys.*`
-/// views take no lock. Nested views (a view over a view) are not expanded here
-/// (they error at query time), so they contribute no locks; view expansion is
-/// one level deep, matching the executor.
-fn read_lock_object_ids(storage: &Storage, db_id: u32, name: &str) -> Vec<u32> {
-    let mut out = Vec::new();
-    let mut visited = std::collections::HashSet::new();
-    collect_read_lock_ids(storage, db_id, name, 0, &mut out, &mut visited);
-    out
-}
-
-/// Adds the locks the AFTER-`event` trigger bodies of `parent_object_id` take,
-/// so a DML that fires them holds every lock its bodies acquire UP FRONT (strict
-/// 2PL — a trigger body reading/writing another table with no pre-acquired lock
-/// is the recurring seam-defect class). Each body is analyzed by the SAME
-/// machinery the batch uses (`analyze_statements_locks`), so its EXEC, TRY/CATCH,
-/// FK-integrity reads, subquery reads, and its own nested triggers are ALL
-/// covered — not a hand-rolled subset. Recursion (a trigger whose body DMLs a
-/// table with its own triggers) is bounded by `trigger_visited` (trigger
-/// object_ids), so a trigger cycle terminates cleanly rather than hanging
-/// analysis under the scheduler mutex.
-#[allow(clippy::too_many_arguments)]
-fn add_trigger_locks(
-    db_id: u32,
-    storage: &Storage,
-    parent_object_id: u32,
-    event: catalog::TriggerEvent,
-    isolation: Isolation,
-    visited: &mut std::collections::HashSet<(String, Isolation)>,
-    trigger_visited: &mut std::collections::HashSet<(u32, Isolation)>,
-    add: &mut impl FnMut(Resource, LockMode),
-) {
-    for trig in storage.rel_triggers_for(parent_object_id, event) {
-        if !trigger_visited.insert((trig.object_id, isolation)) {
-            continue;
-        }
-        let Some(t) = &trig.trigger else { continue };
-        if let Ok(statements) = truthdb_sql::parse_procedure_body(&t.body) {
-            for (resource, mode) in analyze_statements_locks(
-                storage,
-                db_id,
-                &statements,
-                isolation,
-                visited,
-                trigger_visited,
-            ) {
-                add(resource, mode);
-            }
-        }
-    }
-}
-
-/// The base-table object ids the scalar functions called in a SELECT read
-/// through their bodies (deduped). Non-function names collected along the way
-/// resolve to nothing.
-fn select_function_read_ids(storage: &Storage, db_id: u32, select: &Select) -> Vec<u32> {
-    let mut tables = Vec::new();
-    let mut funcs = Vec::new();
-    collect_select_read_names(select, &mut tables, &mut funcs);
-    let mut out = Vec::new();
-    let mut visited = std::collections::HashSet::new();
-    for func in &funcs {
-        collect_read_lock_ids(storage, db_id, func, 0, &mut out, &mut visited);
-    }
-    out
-}
-
-/// Like [`select_function_read_ids`] but for a bare expression (an IF/WHILE
-/// condition).
-fn expr_function_read_ids(storage: &Storage, db_id: u32, expr: &Expr) -> Vec<u32> {
-    let mut tables = Vec::new();
-    let mut funcs = Vec::new();
-    collect_expr_read_names(expr, &mut tables, &mut funcs);
-    let mut out = Vec::new();
-    let mut visited = std::collections::HashSet::new();
-    for func in &funcs {
-        collect_read_lock_ids(storage, db_id, func, 0, &mut out, &mut visited);
-    }
-    out
-}
-
-/// Resolves `name` to the base-table object ids the executor will read,
-/// recursing through nested views (so a view over a view locks the inner view's
-/// base tables). Bounded by [`MAX_VIEW_NESTING`] so a view cycle terminates.
-fn collect_read_lock_ids(
-    storage: &Storage,
-    db_id: u32,
-    name: &str,
-    depth: u32,
-    out: &mut Vec<u32>,
-    visited: &mut std::collections::HashSet<u32>,
-) {
-    if depth > MAX_VIEW_NESTING || name.to_ascii_lowercase().starts_with("sys.") {
-        return;
-    }
-    let Some(def) = resolve_table(storage, db_id, name) else {
-        return;
-    };
-    // Expand each function/view body at most once per analysis. The depth guard
-    // bounds recursion depth but NOT fan-out: a self- or mutually-referential
-    // body that references itself twice would otherwise recurse exponentially
-    // (2^depth), hanging analysis — and, because analyze_locks runs under the
-    // scheduler mutex, freezing every session.
-    if (def.is_function() || def.view_query.is_some()) && !visited.insert(def.object_id) {
-        return;
-    }
-    // A function: its inner reads (subqueries in a scalar body, or an inline
-    // TVF's body SELECT, plus nested function calls) must be locked up front, or
-    // the body would read tables with no lock held under 2PL — the seam-defect
-    // class. Recurse into the body's read targets, bounded by the same guard.
-    if let Some(function) = &def.function {
-        let mut tables = Vec::new();
-        let mut funcs = Vec::new();
-        match &function.returns {
-            FunctionReturns::Scalar { body, .. } => {
-                if let Ok(statements) = truthdb_sql::parse_function_body(body) {
-                    for statement in &statements {
-                        collect_statement_read_names(statement, &mut tables, &mut funcs);
-                    }
-                }
-            }
-            FunctionReturns::InlineTable { select_text } => {
-                if let Ok(body) = parse_view_query(select_text, &def.name) {
-                    let expanded = expand_ctes(&body);
-                    collect_select_read_names(&expanded, &mut tables, &mut funcs);
-                }
-            }
-            // A multi-statement TVF body may read real tables (e.g. INSERT @t
-            // SELECT FROM base): those reads must be locked up front, exactly
-            // like a scalar body. (@-targets are session-local and are skipped
-            // by the read-name collectors, so they add no lock.)
-            FunctionReturns::MultiStatementTable { body, .. } => {
-                if let Ok(statements) = truthdb_sql::parse_table_function_body(body) {
-                    for statement in &statements {
-                        collect_statement_read_names(statement, &mut tables, &mut funcs);
-                    }
-                }
-            }
-        }
-        for referenced in tables.iter().chain(funcs.iter()) {
-            // The body's unqualified names are the FUNCTION's database's.
-            collect_read_lock_ids(
-                storage,
-                def.database_id,
-                referenced,
-                depth + 1,
-                out,
-                visited,
-            );
-        }
-        return;
-    }
-    let Some(text) = &def.view_query else {
-        // A base table.
-        if !out.contains(&def.object_id) {
-            out.push(def.object_id);
-        }
-        return;
-    };
-    // A view: recurse into every table its body references — and every scalar
-    // function it calls, whose body may read further tables (else a UDF reached
-    // through a view would read unlocked). Inline the body's own CTEs so a base
-    // table reached only through a CTE is still locked.
-    let Ok(body) = parse_view_query(text, &def.name) else {
-        return;
-    };
-    let expanded = expand_ctes(&body);
-    let mut tables = Vec::new();
-    let mut funcs = Vec::new();
-    collect_select_read_names(&expanded, &mut tables, &mut funcs);
-    for referenced in tables.iter().chain(funcs.iter()) {
-        // The body's unqualified names are the VIEW's database's.
-        collect_read_lock_ids(
-            storage,
-            def.database_id,
-            referenced,
-            depth + 1,
-            out,
-            visited,
-        );
-    }
-}
-
-/// Views are read-only here; INSERT/UPDATE/DELETE against one is rejected —
-/// and a PROCEDURE is not a data object at all (SQL Server 2809).
-fn reject_dml_on_view(def: &TableDef) -> Result<(), SqlError> {
-    if def.is_procedure() {
-        return Err(procedure_not_a_table(&def.name));
-    }
-    if def.is_function() {
-        return Err(function_not_a_table(&def.name));
-    }
-    if def.is_trigger() {
-        return Err(SqlError::invalid_object(&def.name));
-    }
-    if def.is_view() {
-        return Err(SqlError::new(
-            4406,
-            16,
-            1,
-            format!(
-                "Update or insert of view '{}' is not supported (the view is read-only).",
-                def.name
-            ),
-        ));
-    }
-    Ok(())
-}
-
-/// SQL Server 2809: a procedure referenced where a table/view is required.
-fn procedure_not_a_table(name: &str) -> SqlError {
-    SqlError::new(
-        2809,
-        16,
-        1,
-        format!(
-            "The request for procedure '{name}' failed because '{name}' is a procedure object."
-        ),
-    )
-}
-
-/// A scalar function used where a table is required (`FROM`, DML target,
-/// table-only DDL). A scalar function is not a rowset; SQL Server 4121-class.
-fn function_not_a_table(name: &str) -> SqlError {
-    SqlError::new(
-        4121,
-        16,
-        1,
-        format!(
-            "Cannot find the user-defined function '{name}', or the name refers to a scalar \
-             function that cannot be used where a table is expected."
-        ),
-    )
-}
-
-/// Table-only DDL (ALTER TABLE, CREATE INDEX) rejects a view. Without this a
-/// view's `root_page = 0` would be heap-scanned — and page 0 is the catalog
-/// root, so a bare `ALTER TABLE view ADD CHECK (1=1)` could corrupt the catalog.
-fn reject_view_as_table(def: &TableDef) -> Result<(), SqlError> {
-    if def.is_procedure() {
-        return Err(procedure_not_a_table(&def.name));
-    }
-    if def.is_function() {
-        return Err(function_not_a_table(&def.name));
-    }
-    if def.is_trigger() {
-        return Err(SqlError::invalid_object(&def.name));
-    }
-    if def.is_view() {
-        return Err(SqlError::new(
-            4928,
-            16,
-            1,
-            format!(
-                "Cannot perform this operation on '{}' because it is a view, not a table.",
-                def.name
-            ),
-        ));
-    }
-    Ok(())
-}
-
-/// Validates a CREATE'd object name: one part, or `dbo.<name>`. A database
-/// prefix is refused (SQL Server 166 — CREATE resolves in the current
-/// database only) and an unknown schema is refused (2760). Returns the bare
-/// name to store.
-fn create_object_name<'a>(kind: &str, name: &'a Name) -> Result<&'a str, SqlError> {
-    // A quoted identifier (`[sys.tables]`) is one name, dots and all — the
-    // parser records quoting for the first part, and splitting it here would
-    // invent a schema the user never wrote.
-    if name.quoted {
-        return Ok(strip_schema(&name.value));
-    }
-    let parts: Vec<&str> = name.value.split('.').collect();
-    match parts[..] {
-        [bare] => Ok(bare),
-        [schema, bare] if schema.eq_ignore_ascii_case("dbo") => Ok(bare),
-        [schema, _] => Err(SqlError::new(
-            2760,
-            16,
-            1,
-            format!(
-                "The specified schema name \"{schema}\" either does not exist or you do not have permission to use it."
-            ),
-        )
-        .at(name.span)),
-        _ => Err(SqlError::new(
-            166,
-            15,
-            1,
-            format!(
-                "'{kind}' does not allow specifying the database name as a prefix to the object name."
-            ),
-        )
-        .at(name.span)),
-    }
-}
-
-fn resolve_table(storage: &Storage, db_id: u32, name: &str) -> Option<TableDef> {
-    // Three-part names (`db.dbo.t`, `db..t`) resolve in the named database;
-    // one- and two-part names in the session's. An unknown database or a
-    // schema other than dbo resolves to nothing (208 at the call sites).
-    let parts: Vec<&str> = name.split('.').collect();
-    let (target_db, bare_owned);
-    let bare: &str = match parts[..] {
-        // `[dbo].[my.table]` flattens to dbo.my.table: a leading dbo is the
-        // schema ('dbo' is a reserved database name, so this cannot shadow a
-        // real database) and the REMAINDER is one name containing dots.
-        [schema, ..] if schema.eq_ignore_ascii_case("dbo") => {
-            target_db = db_id;
-            &name[schema.len() + 1..]
-        }
-        [db, schema, t] if schema.is_empty() || schema.eq_ignore_ascii_case("dbo") => {
-            target_db = storage.rel_database_id_by_name(db)?;
-            t
-        }
-        [_, _, _] => return None,
-        _ => {
-            target_db = db_id;
-            bare_owned = strip_schema(name);
-            bare_owned
-        }
-    };
-    if let Some(def) = storage.rel_table(target_db, bare) {
-        return Some(def);
-    }
-    storage
-        .rel_tables()
-        .into_iter()
-        .find(|d| d.database_id == target_db && d.name.eq_ignore_ascii_case(bare))
-}
-
-/// Maps a storage error to a SQL Server-numbered error. PK and NULL
-/// violations are recognized by their storage messages.
-fn map_storage_err(err: StorageError, table: &str) -> SqlError {
-    match err {
-        StorageError::Constraint(msg) if msg.contains("duplicate primary key") => {
-            SqlError::pk_violation(table)
-        }
-        StorageError::Constraint(msg) if msg.contains("duplicate unique index") => {
-            // 2601: cannot insert a duplicate key row in a unique index.
-            SqlError::new(2601, 14, 1, msg)
-        }
-        StorageError::Constraint(msg) if msg.contains("already exists") => {
-            // 1913: an index with that name already exists on the table.
-            SqlError::new(1913, 16, 1, msg)
-        }
-        StorageError::Constraint(msg) if msg.contains("does not allow NULL") => {
-            SqlError::new(515, 16, 2, msg)
-        }
-        StorageError::Constraint(msg) => SqlError::new(547, 16, 0, msg),
-        StorageError::SnapshotSchemaChange(name) => SqlError::new(
-            3961,
-            16,
-            1,
-            format!(
-                "Snapshot isolation transaction failed in database because the object accessed \
-                 by the statement has been modified by a DDL statement in another concurrent \
-                 transaction since the start of this transaction. It is disallowed because the \
-                 metadata is not versioned. Object: '{name}'."
-            ),
-        ),
-        StorageError::InvalidConfig(msg) => SqlError::new(1701, 16, 1, msg),
-        // The WAL ring is full — under FULL recovery this is typically because
-        // un-backed-up log pins truncation (run BACKUP LOG); under SIMPLE it is
-        // an oversized active transaction. SQL Server reports 9002 either way.
-        StorageError::WalFull(msg) => SqlError::new(
-            9002,
-            17,
-            2,
-            format!("The transaction log for the database is full. {msg}"),
-        ),
-        other => SqlError::new(
-            3621,
-            16,
-            1,
-            format!("The statement has been terminated. {other}"),
-        ),
-    }
-}
-
-pub use ast::Statement as SqlStatement;
-
-/// Renders a result cell to its display string (`None` = NULL). Shared by
-/// the JSON envelope and any text renderer.
-pub fn render_cell(datum: &Datum, column_type: &ColumnType) -> Option<String> {
-    value::display(datum, column_type)
 }
